requirements.txt
openai
langchain
pydantic
playwright
load_dotenv
google-search-results